backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.25
python-dotenv==1.0.0
bcrypt==4.1.2
pyjwt==2.8.0
transformers==4.38.0
torch==2.2.0
sendgrid==6.11.0
google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
requests==2.31.0
sendgrid==6.11.0
google-api-python-client==2.111.0
google-auth==2.25.2
requests==2.31.0
